+++ packages/forms/src/UIForm/UIForm.component.js
@@ -1,138 +1,189 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import classNames from 'classnames';
-import tv4 from 'tv4';
+import React, { createRef } from 'react';
 import { translate } from 'react-i18next';
+import classnames from 'classnames';
+import tv4 from 'tv4';
+import memoizeOne from 'memoize-one';
+import omit from 'lodash/omit';
 
-import { DefaultFormTemplate, TextModeFormTemplate } from './FormTemplate';
+import customFormats from './customFormats';
+import getLanguage from './lang';
 import merge from './merge';
-import { formPropTypes } from './utils/propTypes';
-import { validateSingle, validateAll } from './utils/validation';
+import formTemplates from './utils/formTemplates';
+import UIFormPropTypes from './UIForm.propTypes';
+import { UIFormContext } from './hooks/useUIForm';
+import { I18N_DOMAIN_FORMS } from '../constants';
+
+import theme from './UIForm.scss';
 import Widget from './Widget';
 import Buttons from './fields/Button/Buttons.component';
+import defaultWidgets from './utils/widgets';
+import defaultTemplates from './utils/templates';
 import { getValue, mutateValue } from './utils/properties';
-import { removeError, addError } from './utils/errors';
-import getLanguage from './lang';
-import customFormats from './customFormats';
-import { I18N_DOMAIN_FORMS } from '../constants';
-import '../translate';
-import theme from './UIForm.scss';
+import { validateAll, validateSingle } from './utils/validation';
+import { reconciliateAllErrors, reconciliateSingleErrors } from './utils/errors';
 
-export class UIFormComponent extends React.Component {
-	static displayName = 'TalendUIForm';
+function addErrorObject(data) {
+	if (!data.errors) {
+		return { errors: {}, ...data };
+	}
+	return data;
+}
+
+const mergeWidgets = memoizeOne((compatWidgets, userWigets) => ({
+	...defaultWidgets,
+	...compatWidgets,
+	...userWigets,
+}));
+const mergeTemplates = memoizeOne(userTemplates => ({
+	...defaultTemplates,
+	...userTemplates,
+}));
+const mergeSchema = memoizeOne(merge);
+
+class UIForm extends React.Component {
 	constructor(props) {
 		super(props);
-		const { jsonSchema, uiSchema } = props;
+		const { jsonSchema, uiSchema } = this.props.data || this.props.initialData;
 
-		const state = {};
+		let widgets;
+		let mergedSchema;
 		if (Object.keys(jsonSchema).length) {
-			Object.assign(state, merge(jsonSchema, uiSchema));
+			const mergeResult = merge(jsonSchema, uiSchema);
+			mergedSchema = mergeResult.mergedSchema;
+			widgets = mergeWidgets(mergeResult.widgets, props.widgets);
 		}
-		state.widgets = { ...state.widgets, ...props.widgets };
-		this.state = state;
+		this.state = {
+			initial: addErrorObject(this.props.initialData || this.props.data),
+			live: addErrorObject(this.props.data || this.props.initialData),
+			templates: mergeTemplates(props.templates),
+			mergedSchema,
+			widgets,
+		};
+		this.formRef = createRef();
+		this.initLanguage();
 
 		this.onChange = this.onChange.bind(this);
 		this.onFinish = this.onFinish.bind(this);
-		this.onSubmit = this.onSubmit.bind(this);
 		this.onTrigger = this.onTrigger.bind(this);
-		this.onActionClick = this.onActionClick.bind(this);
-		this.focusFirstError = this.focusFirstError.bind(this);
-		this.setFormRef = this.setFormRef.bind(this);
+		this.onReset = this.onReset.bind(this);
+		this.onSubmit = this.onSubmit.bind(this);
+		this.onErrors = this.onErrors.bind(this);
+
+		this.renderWidgets = this.renderWidgets.bind(this);
+		this.renderButtons = this.renderButtons.bind(this);
+	}
+
+	componentWillReceiveProps(nextProps) {
+		const nextDerivedState = {};
+
+		if (nextProps.data !== this.props.data) {
+			nextDerivedState.live = addErrorObject(nextProps.data);
+		}
+		if (nextProps.initialData !== this.props.initialData) {
+			nextDerivedState.initial = addErrorObject(nextProps.initialData);
+			nextDerivedState.live = nextDerivedState.initial;
+		}
+
+		if (nextDerivedState.live) {
+			const { mergedSchema = [], widgets: compatWidgets } = mergeSchema(
+				nextDerivedState.live.jsonSchema,
+				nextDerivedState.live.uiSchema,
+			);
+			nextDerivedState.mergedSchema = mergedSchema;
+			nextDerivedState.widgets = mergeWidgets(compatWidgets, nextProps.widgets);
+		}
+
+		return Object.keys(nextDerivedState).length ? nextDerivedState : null;
+	}
+
+	// eslint-disable-next-line react/sort-comp
+	initLanguage() {
+		const { language, customFormats: userCustomFormats, t } = this.props;
+
 		// control the tv4 language here.
-		const language = getLanguage(props.t);
-		if (props.language != null) {
-			Object.assign(language, props.language);
+		const mergedLanguage = getLanguage(t);
+		if (language != null) {
+			Object.assign(mergedLanguage, language);
 			// Force update of language @talend even if already set
-			tv4.addLanguage('@talend', language);
+			tv4.addLanguage('@talend', mergedLanguage);
 			tv4.language('@talend');
 		}
 		if (!tv4.language('@talend')) {
-			tv4.addLanguage('@talend', language);
+			tv4.addLanguage('@talend', mergedLanguage);
 			tv4.language('@talend'); // set it
 		}
-		const allFormats = Object.assign(customFormats(props.t), props.customFormats);
+		const allFormats = Object.assign(customFormats(t), userCustomFormats);
 		tv4.addFormat(allFormats);
 	}
 
-	/**
-	 * Update the state with the new schema.
-	 * @param jsonSchema
-	 * @param uiSchema
-	 */
-	componentWillReceiveProps({ jsonSchema, uiSchema }) {
-		if (
-			!jsonSchema ||
-			!uiSchema ||
-			(this.props.jsonSchema === jsonSchema && this.props.uiSchema === uiSchema)
-		) {
-			return;
-		}
-		if (Object.keys(jsonSchema).length) {
-			const merged = merge(jsonSchema, uiSchema);
-			this.setState({
-				...merged,
-				widgets: {
-					...merged.widgets,
-					...this.props.widgets,
+	onErrors(event, newErrors, origin) {
+		this.setState(
+			oldState => ({
+				...oldState,
+				live: {
+					...oldState.live,
+					errors: typeof newErrors === 'function' ? newErrors(oldState.live.errors) : newErrors,
 				},
-			});
-		}
+			}),
+			() => {
+				if (origin !== 'submit') {
+					return;
+				}
+				const elementWithError = this.formRef.current.querySelector('[aria-invalid="true"]');
+				if (elementWithError) {
+					elementWithError.focus();
+				}
+			},
+		);
 	}
 
-	/**
-	 * Fire onChange callback while interacting with form fields
-	 * @param event The event that triggered the callback
-	 * @param schema The payload field schema
-	 * @param value The payload new value
-	 */
+	handleNewProperties(event, { schema, value, properties }) {
+		const oldProperties = this.state.live.properties;
+		this.setState(
+			oldState => ({
+				...oldState,
+				live: {
+					...oldState.live,
+					properties:
+						typeof properties === 'function' ? properties(oldState.live.properties) : properties,
+				},
+			}),
+			() => {
+				if (!this.props.onChange) {
+					return;
+				}
+				this.props.onChange(event, {
+					schema,
+					value,
+					oldProperties,
+					properties: this.state.live.properties,
+					formData: this.state.live.properties,
+				});
+			},
+		);
+	}
 	onChange(event, { schema, value }) {
-		const newProperties = mutateValue(this.props.properties, schema, value);
-		if (newProperties !== this.props.properties) {
-			this.props.onChange(event, {
-				schema,
-				value,
-				oldProperties: this.props.properties,
-				properties: newProperties,
-				formData: newProperties,
-			});
-		}
+		this.handleNewProperties(event, {
+			schema,
+			value,
+			properties: oldProperties => mutateValue(oldProperties, schema, value),
+		});
 	}
 
-	/**
-	 * Perform validation and triggers when user has finished to change a value
-	 * @param event The event that triggered the callback
-	 * @param schema The payload field schema
-	 * @param value The new value, provided if not already taken into account
-	 * @param deepValidation Validate the subItems
-	 * Most of the time, this value is not provided. It will be taken from props.properties
-	 * This allows to perform triggers/validation while changing a value
-	 * (ex: add an element in array)
-	 * @param widgetChangeErrors Change errors hook, allows any widget to manipulate the errors map
-	 * (ex: shift the errors in array elements on remove)
-	 */
 	onFinish(event, { schema, value }, { deepValidation = false, widgetChangeErrors } = {}) {
 		// get property value
-		let newValue;
-		if (value !== undefined) {
-			newValue = value;
-		} else {
-			newValue = getValue(this.props.properties, schema);
-		}
+		const newValue = value === undefined ? getValue(this.state.live.properties, schema) : value;
+
 		// validate value. This validation can be deep if schema is an object or an array
 		const widgetErrors = validateSingle(
 			schema,
 			newValue,
-			this.props.properties,
+			this.state.live.properties,
 			this.props.customValidation,
 			deepValidation,
 		);
 		const hasErrors = Object.values(widgetErrors).find(Boolean);
-
-		// update errors map
-		let errors = Object.entries(widgetErrors).reduce((accu, [errorKey, errorValue]) => {
-			const errorSchema = { key: errorKey };
-			return errorValue ? addError(accu, errorSchema, errorValue) : removeError(accu, errorSchema);
-		}, this.props.errors);
+		let errors = reconciliateSingleErrors(this.state.live.errors, widgetErrors);
 
 		// widget error modifier
 		if (widgetChangeErrors) {
@@ -140,12 +191,14 @@ export class UIFormComponent extends React.Component {
 		}
 
 		// commit errors
-		if (errors !== this.props.errors) {
-			this.props.setErrors(event, errors);
+		if (errors !== this.state.live.errors) {
+			this.onErrors(event, errors, 'finish');
 		}
 
+		// legacy compatibility: trigger "after" is determined by onEvent = undefined
+		// and is run on widget finish
 		if (!hasErrors && schema.triggers && schema.triggers.length) {
-			let formData = this.props.properties;
+			let formData = this.state.live.properties;
 			if (value !== undefined) {
 				formData = mutateValue(formData, schema, value);
 			}
@@ -159,7 +212,7 @@ export class UIFormComponent extends React.Component {
 				propertyName = schema.key[schema.key.length - 1];
 				this.onTrigger(event, { formData, formId: this.props.id, propertyName, value });
 			} else {
-				const trigger = schema.triggers.find(t => t.onEvent === undefined);
+				const trigger = schema.triggers.find(trig => trig.onEvent === undefined);
 				if (trigger) {
 					this.onTrigger(event, {
 						trigger,
@@ -173,240 +226,219 @@ export class UIFormComponent extends React.Component {
 		}
 	}
 
-	/**
-	 * Triggers an onTrigger callback that is allowed to modify the form
-	 * @param event The event that triggered the callback
-	 * @param payload The trigger payload
-	 * trigger The type of trigger
-	 * schema The field schema
-	 */
 	onTrigger(event, payload) {
-		const { onTrigger } = this.props;
-		if (!onTrigger) {
+		if (!this.props.onTrigger) {
 			return null;
 		}
 
+		const updateUIFormData = triggerResult => {
+			const { errors, properties } = triggerResult;
+			if (errors) {
+				this.onErrors(event, errors, 'trigger');
+			}
+			if (properties) {
+				this.handleNewProperties(event, { ...payload, properties });
+			}
+			return triggerResult;
+		};
+
 		if (this.props.moz) {
-			return onTrigger(payload.formData, payload.formId, payload.propertyName, payload.value);
+			this.props
+				.onTrigger(payload.formData, payload.formId, payload.propertyName, payload.value)
+				.then(updateUIFormData);
 		}
+
 		if (!payload.trigger) {
 			throw new Error('onTrigger payload do not have required trigger property');
 		}
-		return onTrigger(event, {
-			properties: this.props.properties,
-			errors: this.props.errors,
-			...payload,
-		});
+		return this.props
+			.onTrigger(event, {
+				properties: this.state.live.properties,
+				errors: this.state.live.errors,
+				...payload,
+			})
+			.then(updateUIFormData);
 	}
 
-	onActionClick(actionOnClick) {
-		if (typeof actionOnClick === 'function') {
-			return (event, data) =>
-				actionOnClick(event, {
-					...data,
-					formData: this.props.properties,
-					properties: this.props.properties,
-				});
+	onReset(event) {
+		this.setState(oldState => ({
+			...oldState,
+			live: oldState.initial,
+		}));
+		if (this.props.onReset) {
+			this.props.onReset(event);
 		}
-		return () => {};
 	}
 
-	/**
-	 * Triggers submit callback if form is valid
-	 * @param event the submit event
-	 */
 	onSubmit(event) {
 		if (this.props.onSubmit) {
 			event.preventDefault();
 		}
 
-		const { mergedSchema } = this.state;
-		const { properties, customValidation } = this.props;
-		const newErrors = validateAll(mergedSchema, properties, customValidation);
-		Object.entries(this.props.errors)
-			.filter(entry => entry[0] in newErrors)
-			.reduce((accu, [key, value]) => {
-				// eslint-disable-next-line no-param-reassign
-				accu[key] = value;
-				return accu;
-			}, newErrors);
-
-		const errors = Object.entries(newErrors)
-			.filter(entry => entry[1])
-			.reduce((accu, [key, value]) => {
-				// eslint-disable-next-line no-param-reassign
-				accu[key] = value;
-				return accu;
-			}, {});
-
+		const newErrors = validateAll(
+			this.state.mergedSchema,
+			this.state.live.properties,
+			this.props.customValidation,
+		);
+		const errors = reconciliateAllErrors(this.state.live.errors, newErrors);
 		const isValid = !Object.keys(errors).length;
-		this.props.setErrors(event, errors, !isValid ? this.focusFirstError : undefined);
-
-		if (this.props.onSubmit && isValid) {
-			if (this.props.moz) {
-				this.props.onSubmit(null, { formData: properties });
-			} else {
-				this.props.onSubmit(event, properties, mergedSchema);
+		if (isValid) {
+			this.setState(oldState => ({
+				...oldState,
+				initial: oldState.live,
+			}));
+			if (this.props.onSubmit) {
+				if (this.props.moz) {
+					this.props.onSubmit(null, { formData: this.state.live.properties });
+				} else {
+					this.props.onSubmit(event, this.state.live.properties, this.state.mergedSchema);
+				}
 			}
+		} else {
+			this.onErrors(event, errors, 'submit');
 		}
 
 		return isValid;
 	}
 
-	setFormRef(element) {
-		this.formRef = element;
+	renderWidgets() {
+		return this.state.mergedSchema.map((nextSchema, index) => (
+			<Widget key={nextSchema.key || index} schema={nextSchema} />
+		));
 	}
 
-	focusFirstError() {
-		if (!this.formRef) {
-			return;
-		}
-
-		const elementWithError = this.formRef.querySelector('[aria-invalid="true"]');
-
-		if (elementWithError) {
-			elementWithError.focus();
-		}
-	}
-
-	render() {
-		const { onSubmitEnter, onSubmitLeave, properties } = this.props;
-		const actions = this.props.actions || [
+	renderButtons() {
+		const {
+			actions,
+			buttonBlockClass = 'form-actions',
+			getComponent,
+			id,
+			onSubmitEnter,
+			onSubmitLeave,
+		} = this.props;
+		const buttonsSchemas = actions || [
 			{
 				bsStyle: 'primary',
 				label: 'Submit',
 				type: 'submit',
 				widget: 'button',
 				position: 'right',
-				onMouseEnter: onSubmitEnter && (event => onSubmitEnter(event, properties)),
+				onMouseEnter: onSubmitEnter && (event => onSubmitEnter(event, this.state.live.properties)),
 				onMouseLeave: onSubmitLeave,
 			},
 		];
+		if (buttonsSchemas.length === 0) {
+			return null;
+		}
+
+		// TODO move handleActionClick to Buttons and use context to get properties
+		function handleActionClick(actionOnClick) {
+			if (typeof actionOnClick === 'function') {
+				return (event, payload) =>
+					actionOnClick(event, {
+						...payload,
+						formData: this.state.live.properties,
+						properties: this.state.live.properties,
+					});
+			}
+			return () => {};
+		}
+		return (
+			<div className={classnames(theme['form-actions'], 'tf-actions-wrapper')} key="form-buttons">
+				<Buttons
+					id={`${id}-actions`}
+					onTrigger={this.onTrigger}
+					className={buttonBlockClass}
+					schema={{ items: buttonsSchemas }}
+					onClick={handleActionClick}
+					getComponent={getComponent}
+				/>
+			</div>
+		);
+	}
+
+	render() {
 		if (!this.state.mergedSchema) {
 			return null;
 		}
 
-		const formTemplate =
-			this.props.displayMode === 'text' ? TextModeFormTemplate : DefaultFormTemplate;
-		const widgetsRenderer = () =>
-			this.state.mergedSchema.map((nextSchema, index) => (
-				<Widget
-					id={this.props.id}
-					key={index}
-					onChange={this.onChange}
-					onFinish={this.onFinish}
-					onTrigger={this.onTrigger}
-					schema={nextSchema}
-					properties={this.props.properties}
-					errors={this.props.errors}
-					templates={this.props.templates}
-					widgets={this.state.widgets}
-					displayMode={this.props.displayMode}
-					updating={this.props.updating}
-				/>
-			));
-		const buttonsRenderer = () => {
-			if (actions.length === 0) {
-				return null;
-			}
-
-			return (
-				<div className={classNames(theme['form-actions'], 'tf-actions-wrapper')} key="form-buttons">
-					<Buttons
-						id={`${this.props.id}-${this.props.id}-actions`}
-						onTrigger={this.onTrigger}
-						className={this.props.buttonBlockClass}
-						schema={{ items: actions }}
-						onClick={this.onActionClick}
-						getComponent={this.props.getComponent}
-					/>
-				</div>
-			);
+		const {
+			children,
+			className,
+			displayMode,
+			id,
+			noHtml5Validate,
+			noValidate,
+			updating,
+		} = this.props;
+		const formProps = {
+			...omit(this.props, [
+				'actions',
+				'buttonBlockClass',
+				'children',
+				'className',
+				'customFormats',
+				'customValidation',
+				'data',
+				'displayMode',
+				'getComponent',
+				'i18n',
+				'initialData',
+				'language',
+				'moz',
+				'onChange',
+				'onErrors',
+				'onReset',
+				'onSubmit',
+				'onSubmitEnter',
+				'onSubmitLeave',
+				'onTrigger',
+				't',
+				'tReady',
+				'templates',
+				'widgets',
+			]),
+			className: classnames('tf-uiform', theme.uiform, className),
+			noValidate: noValidate || noHtml5Validate,
+			onReset: this.onReset,
+			onSubmit: this.onSubmit,
 		};
 
+		const contextValue = {
+			displayMode,
+			id,
+			onChange: this.onChange,
+			onFinish: this.onFinish,
+			onTrigger: this.onTrigger,
+			state: {
+				errors: this.state.live.errors,
+				properties: this.state.live.properties,
+			},
+			templates: this.state.templates,
+			widgets: this.state.widgets,
+			updating,
+		};
+		const FormTemplate = formTemplates[displayMode] || formTemplates.default;
+
 		return (
-			<form
-				acceptCharset={this.props.acceptCharset}
-				action={this.props.action}
-				autoComplete={this.props.autoComplete}
-				className={classNames('tf-uiform', theme.uiform, this.props.className)}
-				encType={this.props.encType}
-				id={this.props.id}
-				method={this.props.method}
-				name={this.props.name}
-				noValidate={this.props.noHtml5Validate}
-				onReset={this.props.onReset}
-				onSubmit={this.onSubmit}
-				target={this.props.target}
-				ref={this.setFormRef}
-			>
-				{formTemplate({ children: this.props.children, widgetsRenderer, buttonsRenderer })}
-			</form>
+			<UIFormContext.Provider value={contextValue}>
+				<FormTemplate
+					formProps={formProps}
+					renderWidgets={this.renderWidgets}
+					renderButtons={this.renderButtons}
+					formRef={this.formRef}
+				>
+					{children}
+				</FormTemplate>
+			</UIFormContext.Provider>
 		);
 	}
 }
-const I18NUIForm = translate(I18N_DOMAIN_FORMS)(UIFormComponent);
+
+UIForm.displayName = 'UIForm';
 
 if (process.env.NODE_ENV !== 'production') {
-	I18NUIForm.propTypes = {
-		...formPropTypes,
-
-		/** Form definition: Json schema that specify the data model */
-		jsonSchema: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
-		/** Form definition: UI schema that specify how to render the fields */
-		uiSchema: PropTypes.array.isRequired, // eslint-disable-line react/forbid-prop-types
-		/**
-		 *  Form definition: Form fields values.
-		 *  Note that it should contains @definitionName for triggers.
-		 */
-		properties: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
-		/** Form definition: The forms errors { [fieldKey]: errorMessage } */
-		errors: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
-
-		/**
-		 * Actions buttons to display at the bottom of the form.
-		 * If not provided, a single submit button is displayed.
-		 */
-		actions: PropTypes.arrayOf(Buttons.propTypes.schema),
-		/**
-		 * User callback: Custom validation function.
-		 * Prototype: function customValidation(schema, value, properties)
-		 * Return format : errorMessage String | falsy
-		 * This is triggered on fields that has their uiSchema > customValidation : true
-		 */
-		customValidation: PropTypes.func,
-		/*
-		 * Form definition: Custom formats
-		 */
-		customFormats: PropTypes.object,
-		/**
-		 * User callback: Trigger
-		 * Prototype: function onTrigger(event, { trigger, schema, properties })
-		 */
-		onTrigger: PropTypes.func,
-		/** Custom templates */
-		templates: PropTypes.object,
-		/** Custom widgets */
-		widgets: PropTypes.object,
-		/** Display mode: example 'text' */
-		displayMode: PropTypes.string,
-
-		/** State management impl: The change callback */
-		onChange: PropTypes.func.isRequired,
-		/** State management impl: Set All fields validations errors */
-		setErrors: PropTypes.func,
-		getComponent: PropTypes.func,
-		onSubmitEnter: PropTypes.func,
-		onSubmitLeave: PropTypes.func,
-	};
-	UIFormComponent.propTypes = I18NUIForm.propTypes;
+	UIForm.propTypes = UIFormPropTypes;
 }
 
-I18NUIForm.defaultProps = {
-	noHtml5Validate: true,
-	buttonBlockClass: 'form-actions',
-	properties: {},
-};
-UIFormComponent.defaultProps = I18NUIForm.defaultProps;
-
-export default I18NUIForm;
+export default translate(I18N_DOMAIN_FORMS)(UIForm);
