+++ packages/forms/src/UIForm/UIForm.propTypes.js
@@ -77,8 +77,10 @@ export default {
 	displayMode: PropTypes.string,
 	/** Translations */
 	language: PropTypes.object,
-	/** Form definition: prevent html 5 validations */
+	/** Form definition: prevent html 5 validations. This one is for compatibility, prefer 'noValidate' */
 	noHtml5Validate: PropTypes.bool,
+	/** Same as noHtml5Validate but in a standard name */
+	noValidate: PropTypes.bool,
 	/**
 	 * The change callback.
 	 * Prototype: function onChange(event, { schema, value, properties })
