+++ packages/forms/src/UIForm/fieldsets/Array/Array.component.js
@@ -1,10 +1,9 @@
+import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
-import React from 'react';
 import Widget from '../../Widget';
-import { shiftArrayErrorsKeys } from '../../utils/validation';
-import defaultTemplates from '../../utils/templates';
-import defaultWidgets from '../../utils/widgets';
+import { UIFormContext } from '../../hooks/useUIForm';
 import { getArrayElementSchema } from '../../utils/array';
+import { shiftArrayErrorsKeys } from '../../utils/validation';
 
 function getRange(previousIndex, nextIndex) {
 	if (previousIndex < nextIndex) {
@@ -20,56 +19,35 @@ function getRange(previousIndex, nextIndex) {
 	};
 }
 
-export default class ArrayWidget extends React.Component {
-	constructor(props) {
-		super(props);
+export default function ArrayWidget(props) {
+	const { id, schema, value } = props;
+	const { displayMode, onChange, onFinish, templates, widgets } = useContext(UIFormContext);
 
-		this.onAdd = this.onAdd.bind(this);
-		this.onRemove = this.onRemove.bind(this);
-		this.onReorder = this.onReorder.bind(this);
-		this.renderItem = this.renderItem.bind(this);
-	}
+	const canReorder = schema.reorder !== false;
+	const templateId = displayMode ? `array_${displayMode}` : 'array';
+	const ArrayTemplate = templates[templateId];
 
-	onAdd(event) {
-		const arrayMergedSchema = this.props.schema;
-		const defaultValue = arrayMergedSchema.schema.items.type === 'object' ? {} : '';
+	const itemWidget = widgets[schema.itemWidget];
+	const isCloseable = itemWidget && itemWidget.isCloseable;
 
-		let currentValue = this.props.value;
-		if (this.isCloseable()) {
+	function onAdd(event) {
+		const defaultValue = schema.schema.items.type === 'object' ? {} : '';
+
+		let currentValue = value;
+		if (isCloseable) {
 			currentValue = currentValue.map(item => ({ ...item, isClosed: true }));
 		}
-		const value = currentValue.concat(defaultValue);
+		const newValue = currentValue.concat(defaultValue);
 
-		const payload = { schema: arrayMergedSchema, value };
-		this.props.onChange(event, payload);
-		this.props.onFinish(event, payload);
+		const payload = { schema, value: newValue };
+		onChange(event, payload);
+		onFinish(event, payload);
 	}
 
-	onRemove(event, indexToRemove) {
-		const schema = this.props.schema;
-		const value = this.props.value.slice(0);
-		value.splice(indexToRemove, 1);
-
-		// shift up the items errors after the one we remove
-		function widgetChangeErrors(errors) {
-			return shiftArrayErrorsKeys(errors, {
-				arrayKey: schema.key,
-				minIndex: indexToRemove,
-				shouldRemoveIndex: index => index === indexToRemove,
-				getNextIndex: index => index - 1,
-			});
-		}
-
-		const payload = { schema, value };
-		this.props.onChange(event, payload);
-		this.props.onFinish(event, payload, { widgetChangeErrors });
-	}
-
-	onReorder(event, { previousIndex, nextIndex }) {
-		const schema = this.props.schema;
-		const value = this.props.value.slice(0);
-		const [item] = value.splice(previousIndex, 1);
-		value.splice(nextIndex, 0, item);
+	function onReorder(event, { previousIndex, nextIndex }) {
+		const newValue = value.slice(0);
+		const [item] = newValue.splice(previousIndex, 1);
+		newValue.splice(nextIndex, 0, item);
 
 		function widgetChangeErrors(errors) {
 			// determine the range [min, max[ of items to shift, with the pace
@@ -88,78 +66,61 @@ export default class ArrayWidget extends React.Component {
 			});
 		}
 
-		const payload = { schema, value };
-		this.props.onChange(event, payload);
-		this.props.onFinish(event, payload, { widgetChangeErrors });
+		const payload = { schema, value: newValue };
+		onChange(event, payload);
+		onFinish(event, payload, { widgetChangeErrors });
 	}
 
-	getArrayTemplate() {
-		const baseTemplateId = 'array';
-		const templateId = `${baseTemplateId}_${this.props.displayMode}`;
-		const ArrayTemplate = this.props.templates[templateId] || defaultTemplates[templateId];
-		if (!ArrayTemplate) {
-			return this.props.templates[baseTemplateId] || defaultTemplates[baseTemplateId];
+	function onRemove(event, indexToRemove) {
+		const newValue = value.slice(0);
+		newValue.splice(indexToRemove, 1);
+
+		// shift up the items errors after the one we remove
+		function widgetChangeErrors(errors) {
+			return shiftArrayErrorsKeys(errors, {
+				arrayKey: schema.key,
+				minIndex: indexToRemove,
+				shouldRemoveIndex: index => index === indexToRemove,
+				getNextIndex: index => index - 1,
+			});
 		}
-		return ArrayTemplate;
+
+		const payload = { schema, value: newValue };
+		onChange(event, payload);
+		onFinish(event, payload, { widgetChangeErrors });
 	}
 
-	isCloseable() {
-		const widgetId = this.props.schema.itemWidget;
-		const itemWidget = this.props.widgets[widgetId] || defaultWidgets[widgetId];
-		if (!itemWidget) {
-			return false;
-		}
-		return itemWidget.isCloseable === true;
-	}
-
-	renderItem(index, extraProps) {
+	function renderItem(index, extraProps) {
 		return (
 			<Widget
-				{...this.props}
 				{...extraProps}
-				disabled={this.props.schema.disabled}
-				id={this.props.id && `${this.props.id}-${index}`}
-				schema={getArrayElementSchema(this.props.schema, index)}
-				value={this.props.value[index]}
+				id={id && `${id}-${index}`}
+				schema={getArrayElementSchema(schema, index)}
 			/>
 		);
 	}
 
-	render() {
-		const { schema } = this.props;
-		const canReorder = schema.reorder !== false;
-		const ArrayTemplate = this.getArrayTemplate();
-
-		return (
-			<ArrayTemplate
-				{...this.props}
-				canReorder={canReorder}
-				onAdd={this.onAdd}
-				onReorder={this.onReorder}
-				onRemove={this.onRemove}
-				renderItem={this.renderItem}
-				isCloseable={this.isCloseable()}
-			/>
-		);
-	}
+	return (
+		<ArrayTemplate
+			{...props}
+			canReorder={canReorder}
+			onAdd={onAdd}
+			onReorder={onReorder}
+			onRemove={onRemove}
+			renderItem={renderItem}
+			isCloseable={isCloseable}
+		/>
+	);
 }
 
 ArrayWidget.defaultProps = {
-	items: [],
 	value: [],
-	templates: {},
-	widgets: {},
 };
 
 if (process.env.NODE_ENV !== 'production') {
 	ArrayWidget.propTypes = {
-		displayMode: PropTypes.string,
 		id: PropTypes.string,
-		onChange: PropTypes.func.isRequired,
-		onFinish: PropTypes.func.isRequired,
 		schema: PropTypes.object.isRequired,
-		templates: PropTypes.object.isRequired,
 		value: PropTypes.array.isRequired,
-		widgets: PropTypes.object.isRequired,
 	};
 }
