+++ packages/forms/src/UIForm/index.js
@@ -1,5 +1,4 @@
-//import UIForm from './UIFormWithHooks';
-import UIForm from './UIForm';
+import UIForm from './UIForm.component';
 import * as triggers from './utils/triggers';
 
 export { UIForm, triggers };
