+++ packages/forms/src/UIForm/utils/array.js
@@ -119,5 +119,6 @@ export function getArrayElementSchema(arraySchema, elementIndex) {
 		items: getArrayElementItems(arraySchema, elementIndex),
 		widget: arraySchema.itemWidget || 'fieldset',
 		title: arraySchema.itemTitle,
+		disabled: arraySchema.disabled,
 	};
 }
